chore : service 정리, 주석정리
