+++ src/main/java/com/a404/duckonback/config/SecurityConfig.java
@@ -1,8 +1,6 @@
 package com.a404.duckonback.config;
 
-import com.a404.duckonback.Filter.JWTFilter;
-import com.a404.duckonback.service.UserServiceImpl;
-import com.a404.duckonback.util.JWTUtil;
+import com.a404.duckonback.filter.JWTFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -12,8 +10,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import static org.springframework.security.config.Customizer.withDefaults;
 
 @Configuration
 @RequiredArgsConstructor
