+++ src/main/java/com/a404/duckonback/controller/AuthController.java
@@ -18,7 +18,6 @@ import org.springframework.web.bind.annotation.*;
 public class AuthController {
 
     private final AuthService authService;
-    private final UserService userService;
 
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody LoginRequestDTO loginRequest){
@@ -30,12 +29,29 @@ public class AuthController {
 
 
         try {
-            return userService.signup(request);
+            return authService.signup(request);
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("서버 내부 오류가 발생했습니다.");
         }
     }
+
+//    @PostMapping("/refresh")
+//    public ResponseEntity<?> refreshAccessToken(@RequestHeader("Authorization") String refreshTokenHeader) {
+//        // Bearer 제거
+//        if (!refreshTokenHeader.startsWith("Bearer ")) {
+//            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("잘못된 형식의 토큰입니다.");
+//        }
+//
+//        String refreshToken = refreshTokenHeader.substring(7);
+//
+//        try {
+//            String newAccessToken = authService.refreshAccessToken(refreshToken);
+//            return ResponseEntity.ok().body(Map.of("accessToken", newAccessToken));
+//        } catch (Exception e) {
+//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
+//        }
+//    }
 
 
 }
