+++ src/main/java/com/a404/duckonback/service/AuthServiceImpl.java
@@ -2,21 +2,31 @@ package com.a404.duckonback.service;
 
 import com.a404.duckonback.dto.LoginRequestDTO;
 import com.a404.duckonback.dto.LoginResponseDTO;
+import com.a404.duckonback.dto.SignupRequestDTO;
 import com.a404.duckonback.dto.UserDTO;
 import com.a404.duckonback.entity.User;
+import com.a404.duckonback.enums.UserRole;
 import com.a404.duckonback.exception.CustomException;
+import com.a404.duckonback.repository.UserRepository;
 import com.a404.duckonback.util.JWTUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 public class AuthServiceImpl implements AuthService {
 
-    private final UserServiceImpl userServiceImpl;
-    private final ArtistServiceImpl artistServiceImpl;
+    private final UserService userService;
+    private final ArtistService artistService;
+
+    private final UserRepository userRepository;
 
     private final PasswordEncoder passwordEncoder;
     private final JWTUtil jwtUtil;
@@ -36,9 +46,9 @@ public class AuthServiceImpl implements AuthService {
 
         User user = null;
         if (email != null && !email.isBlank()) {
-            user = userServiceImpl.findByEmail(email);
+            user = userService.findByEmail(email);
         } else if (userId != null && !userId.isBlank()) {
-            user = userServiceImpl.findByUserId(userId);
+            user = userService.findByUserId(userId);
         }
 
         if (user == null) {
@@ -62,7 +72,7 @@ public class AuthServiceImpl implements AuthService {
                 .role(user.getRole().name())
                 .language(user.getLanguage())
                 .imgUrl(user.getImgUrl())
-                .artistList(artistServiceImpl.findAllArtistIdByUserUuid(user.getUuid()))
+                .artistList(artistService.findAllArtistIdByUserUuid(user.getUuid()))
                 .build();
 
         return LoginResponseDTO.builder()
@@ -70,5 +80,33 @@ public class AuthServiceImpl implements AuthService {
                 .refreshToken(refreshToken)
                 .user(userDTO)
                 .build();
+    }
+
+    public ResponseEntity<?> signup(SignupRequestDTO dto){
+        MultipartFile file = dto.getProfileImg();
+        String imgUrl = null;
+        if (file != null && !file.isEmpty()) {
+            //이미지 저장
+        }
+
+        User user = User.builder()
+                .uuid(UUID.randomUUID().toString())
+                .email(dto.getEmail())
+                .userId(dto.getUserId())
+                .password(passwordEncoder.encode(dto.getPassword()))
+                .nickname(dto.getNickname())
+                .createdAt(LocalDateTime.now())
+                .role(UserRole.USER)
+                .language(dto.getLanguage())
+                .imgUrl(imgUrl)
+                .build();
+
+        userService.save(user);
+
+        if (dto.getArtistList() != null && !dto.getArtistList().isEmpty()) {
+            artistService.followArtists(user.getUuid(), dto.getArtistList());
+        }
+
+        return ResponseEntity.ok().body("회원가입이 성공적으로 완료되었습니다!");
     }
 }
