+++ src/main/java/com/a404/duckonback/service/UserService.java
@@ -1,11 +1,9 @@
 package com.a404.duckonback.service;
 
-import com.a404.duckonback.dto.SignupRequestDTO;
 import com.a404.duckonback.entity.User;
-import org.springframework.http.ResponseEntity;
 
 public interface UserService {
     User findByEmail(String email);
     User findByUserId(String userId);
-    ResponseEntity<?> signup(SignupRequestDTO dto);
+    void save(User user);
 }
