+++ src/main/java/com/a404/duckonback/service/UserServiceImpl.java
@@ -19,10 +19,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
-    private final PasswordEncoder passwordEncoder;
-
     private final UserRepository userRepository;
-    private final ArtistServiceImpl artistServiceImpl;
 
     public User findByEmail(String email) {
         return userRepository.findByEmail(email);
@@ -32,31 +29,7 @@ public class UserServiceImpl implements UserService {
         return userRepository.findByUserId(userId);
     }
 
-    public ResponseEntity<?> signup(SignupRequestDTO dto){
-        MultipartFile file = dto.getProfileImg();
-        String imgUrl = null;
-        if (file != null && !file.isEmpty()) {
-            //이미지 저장
-        }
+    public void save(User user) { userRepository.save(user);}
 
-        User user = User.builder()
-                .uuid(UUID.randomUUID().toString())
-                .email(dto.getEmail())
-                .userId(dto.getUserId())
-                .password(passwordEncoder.encode(dto.getPassword()))
-                .nickname(dto.getNickname())
-                .createdAt(LocalDateTime.now())
-                .role(UserRole.USER)
-                .language(dto.getLanguage())
-                .imgUrl(imgUrl)
-                .build();
 
-        userRepository.save(user);
-
-        if (dto.getArtistList() != null && !dto.getArtistList().isEmpty()) {
-            artistServiceImpl.followArtists(user.getUuid(), dto.getArtistList());
-        }
-
-        return ResponseEntity.ok().body("회원가입이 성공적으로 완료되었습니다!");
-    }
 }
